Fix bugs in sending response

+++ src/config/db.ts
@@ -22,7 +22,7 @@ export const initDB = async () => {
         vehicle_name VARCHAR(255) NOT NULL,
         type VARCHAR(10) NOT NULL CHECK (type IN ('car', 'bike', 'van', 'SUV')),
         registration_number VARCHAR(50) UNIQUE NOT NULL,
-        daily_rent_price NUMERIC(10,2) NOT NULL CHECK (daily_rent_price > 0),
+        daily_rent_price INT NOT NULL CHECK (daily_rent_price > 0),
         availability_status VARCHAR(20) NOT NULL CHECK (availability_status IN ('available', 'booked'))
         );`);
 
@@ -33,7 +33,7 @@ export const initDB = async () => {
         vehicle_id INT REFERENCES vehicles(id) ON DELETE CASCADE,
         rent_start_date DATE NOT NULL,
         rent_end_date DATE NOT NULL CHECK (rent_end_date > rent_start_date),
-        total_price NUMERIC(10,2) NOT NULL CHECK (total_price > 0),
+        total_price INT NOT NULL CHECK (total_price > 0),
         status VARCHAR(20) NOT NULL CHECK (status IN ('active', 'canceled', 'returned'))
         );`);
 };
